+++ CHANGELOG.md
@@ -10,12 +10,13 @@ Jump to version:
 Additions
 - Added RotateBy to Direction2Helper
 
-Fixes
-- Set default values for InputHandler held and pressed keys to avoid an exception if buttons are held in the very first frame
-
 Improvements
 - Improved NinePatch memory performance
 - Moved sound-related classes into Sound namespace
+- Added customizable overloads for Keybind, Combination and GenericInput ToString methods 
+
+Fixes
+- Set default values for InputHandler held and pressed keys to avoid an exception if buttons are held in the very first frame
 
 ### MLEM.Ui
 Additions